Move foreign influence decisions to their own file.
