+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj.filters
@@ -462,6 +462,9 @@
     <ClCompile Include="Source\HOI4World\Decisions\ExiledGovernmentsDecisions.cpp">
       <Filter>HoI4World\Decisions</Filter>
     </ClCompile>
+    <ClCompile Include="Source\HOI4World\Decisions\ForeignInfluenceDecisions.cpp">
+      <Filter>HoI4World\Decisions</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\common_items\Date.h">
@@ -879,6 +882,9 @@
       <Filter>HoI4World\Decisions</Filter>
     </ClInclude>
     <ClInclude Include="Source\HOI4World\Decisions\ExiledGovernmentsDecisions.h">
+      <Filter>HoI4World\Decisions</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Decisions\ForeignInfluenceDecisions.h">
       <Filter>HoI4World\Decisions</Filter>
     </ClInclude>
   </ItemGroup>
